\name{PeakErrorChrom}
\alias{PeakErrorChrom}
\title{PeakErrorChrom}
\description{Assume p and r come from the same chromosome.}
\usage{PeakErrorChrom(p, r)}
\arguments{
  \item{p}{data.frame of peaks.}
  \item{r}{data.frame of regions.}
}

\value{data.frame of error.}

\author{Toby Dylan Hocking}





