\name{Peaks}
\alias{Peaks}
\title{Peaks}
\description{Represent a data.frame that represents a list of peaks.}
\usage{Peaks(chrom = factor(), chromStart = integer(), chromEnd = integer())}
\arguments{
  \item{chrom}{character or factor with chrom name for example "chr22"}
  \item{chromStart}{numeric, first base of peak.}
  \item{chromEnd}{numeric, base after peak.}
}

\value{data.frame with columns chrom, chromStart, chromEnd.}

\author{Toby Dylan Hocking}





